requirements.txt
-i https://pypi.org/simple
--extra-index-url https://download.pytorch.org/whl/cu121
certifi==2024.2.2; python_version >= '3.6'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
colorama==0.4.6; platform_system == 'Windows'
contourpy==1.2.0; python_version >= '3.9'
cycler==0.12.1; python_version >= '3.8'
easyocr==1.7.1
filelock==3.13.1; python_version >= '3.8'
fonttools==4.49.0; python_version >= '3.8'
fsspec==2024.2.0; python_version >= '3.8'
idna==3.6; python_version >= '3.5'
imageio==2.34.0; python_version >= '3.8'
jinja2==3.1.3; python_version >= '3.7'
kiwisolver==1.4.5; python_version >= '3.7'
lazy-loader==0.3; python_version >= '3.7'
markupsafe==2.1.5; python_version >= '3.7'
matplotlib==3.8.3; python_version >= '3.9'
mpmath==1.3.0
networkx==3.2.1; python_version >= '3.9'
ninja==1.11.1.1
numpy==1.26.4; python_version >= '3.9'
opencv-python==4.9.0.80; python_version >= '3.6'
opencv-python-headless==4.9.0.80; python_version >= '3.6'
packaging==23.2; python_version >= '3.7'
pandas==2.2.0; python_version >= '3.9'
pillow==10.2.0; python_version >= '3.8'
psutil==5.9.8; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
py-cpuinfo==9.0.0
pyclipper==1.3.0.post5
pyparsing==3.1.1; python_full_version >= '3.6.8'
pytesseract==0.3.10; python_version >= '3.7'
python-bidi==0.4.2
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2024.1
pyyaml==6.0.1; python_version >= '3.6'
requests==2.31.0; python_version >= '3.7'
scikit-image==0.22.0; python_version >= '3.9'
scipy==1.12.0; python_version >= '3.9'
seaborn==0.13.2; python_version >= '3.8'
shapely==2.0.3; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sympy==1.12; python_version >= '3.8'
thop==0.1.1.post2209072238
tifffile==2024.2.12; python_version >= '3.9'
torch==2.2.0+cu121
torchaudio==2.2.0+cu121
torchvision==0.17.0+cu121
tqdm==4.66.2; python_version >= '3.7'
typing-extensions==4.9.0; python_version >= '3.8'
tzdata==2024.1; python_version >= '2'
ultralytics==8.1.15; python_version >= '3.8'
urllib3==2.2.1; python_version >= '3.8'
